PostRequest class Utils class
