+++ .idea/workspace.xml
@@ -8,8 +8,12 @@
   </component>
   <component name="ChangeListManager">
     <list default="true" id="9c0b0719-30fb-46eb-8ef4-6a93dfe85f5c" name="Changes" comment="">
-      <change beforePath="$PROJECT_DIR$/.github/workflows/maven.yml" beforeDir="false" afterPath="$PROJECT_DIR$/.github/workflows/maven.yml" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/.idea/misc.xml" beforeDir="false" afterPath="$PROJECT_DIR$/.idea/misc.xml" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/.idea/runConfigurations.xml" beforeDir="false" />
       <change beforePath="$PROJECT_DIR$/.idea/workspace.xml" beforeDir="false" afterPath="$PROJECT_DIR$/.idea/workspace.xml" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/report/report.html" beforeDir="false" afterPath="$PROJECT_DIR$/report/report.html" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/src/main/java/functions/Requests.java" beforeDir="false" afterPath="$PROJECT_DIR$/src/main/java/functions/GetRequests.java" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/src/main/java/hooks/Hooks.java" beforeDir="false" afterPath="$PROJECT_DIR$/src/main/java/hooks/Hooks.java" afterDir="false" />
       <change beforePath="$PROJECT_DIR$/src/test/java/TestsRequest.java" beforeDir="false" afterPath="$PROJECT_DIR$/src/test/java/TestsRequest.java" afterDir="false" />
     </list>
     <option name="SHOW_DIALOG" value="false" />
@@ -25,6 +29,11 @@
     </option>
   </component>
   <component name="Git.Settings">
+    <option name="RECENT_BRANCH_BY_REPOSITORY">
+      <map>
+        <entry key="$PROJECT_DIR$" value="main" />
+      </map>
+    </option>
     <option name="RECENT_GIT_ROOT_PATH" value="$PROJECT_DIR$" />
   </component>
   <component name="GitSEFilterConfiguration">
@@ -38,6 +47,9 @@
   <component name="MarkdownSettingsMigration">
     <option name="stateVersion" value="1" />
   </component>
+  <component name="ProjectColorInfo"><![CDATA[{
+  "associatedIndex": 6
+}]]></component>
   <component name="ProjectId" id="2WaaU2lwf882n9E36GeZSfSRCUc" />
   <component name="ProjectLevelVcsManager" settingsEditedManually="true">
     <ConfirmationsSetting value="1" id="Add" />
@@ -46,16 +58,18 @@
     <option name="hideEmptyMiddlePackages" value="true" />
     <option name="showLibraryContents" value="true" />
   </component>
-  <component name="PropertiesComponent">
-    <property name="RunOnceActivity.OpenProjectViewOnStart" value="true" />
-    <property name="RunOnceActivity.ShowReadmeOnStart" value="true" />
-    <property name="last_opened_file_path" value="$PROJECT_DIR$" />
-    <property name="project.structure.last.edited" value="SDKs" />
-    <property name="project.structure.proportion" value="0.2375" />
-    <property name="project.structure.side.proportion" value="0.4717286" />
-    <property name="settings.editor.selected.configurable" value="preferences.pluginManager" />
-  </component>
-  <component name="RunManager" selected="TestNG.TestsRequest">
+  <component name="PropertiesComponent"><![CDATA[{
+  "keyToString": {
+    "WebServerToolWindowFactoryState": "false",
+    "git-widget-placeholder": "NewBrunosBranch",
+    "node.js.detected.package.eslint": "true",
+    "node.js.detected.package.tslint": "true",
+    "node.js.selected.package.eslint": "(autodetect)",
+    "node.js.selected.package.tslint": "(autodetect)",
+    "vue.rearranger.settings.migration": "true"
+  }
+}]]></component>
+  <component name="RunManager" selected="TestNG.TestsRequest (1)">
     <configuration default="true" type="JetRunConfigurationType">
       <module name="com.br.highlevel.apiAutomationRest" />
       <method v="2">
@@ -69,8 +83,8 @@
         <option name="Make" enabled="true" />
       </method>
     </configuration>
-    <configuration name="TestsRequest" type="TestNG" temporary="true" nameIsGenerated="true">
-      <module name="com.br.highlevel.apiAutomationRest" />
+    <configuration name="TestsRequest (1)" type="TestNG" temporary="true" nameIsGenerated="true">
+      <module name="javarestassuredextentreport" />
       <option name="PACKAGE_NAME" value="" />
       <option name="MAIN_CLASS_NAME" value="TestsRequest" />
       <option name="TEST_OBJECT" value="CLASS" />
@@ -80,12 +94,11 @@
         <option name="Make" enabled="true" />
       </method>
     </configuration>
-    <configuration name="TestsRequest.getAllUsers" type="TestNG" temporary="true" nameIsGenerated="true">
+    <configuration name="TestsRequest" type="TestNG" temporary="true" nameIsGenerated="true">
       <module name="com.br.highlevel.apiAutomationRest" />
       <option name="PACKAGE_NAME" value="" />
       <option name="MAIN_CLASS_NAME" value="TestsRequest" />
-      <option name="METHOD_NAME" value="getAllUsers" />
-      <option name="TEST_OBJECT" value="METHOD" />
+      <option name="TEST_OBJECT" value="CLASS" />
       <properties />
       <listeners />
       <method v="2">
@@ -150,20 +163,20 @@
       </method>
     </configuration>
     <list>
+      <item itemvalue="TestNG.Unnamed" />
+      <item itemvalue="TestNG.TestsRequest (1)" />
       <item itemvalue="TestNG.TestsRequest" />
-      <item itemvalue="TestNG.TestsRequest.getAllUsers" />
       <item itemvalue="TestNG.TestsRequest.getProductsByNonexistentId" />
       <item itemvalue="TestNG.TestsRequest.getProductsWById" />
       <item itemvalue="TestNG.TestsRequest.getProductsWithoutAuth" />
-      <item itemvalue="TestNG.Unnamed" />
     </list>
     <recent_temporary>
       <list>
+        <item itemvalue="TestNG.TestsRequest (1)" />
         <item itemvalue="TestNG.TestsRequest" />
         <item itemvalue="TestNG.TestsRequest.getProductsWithoutAuth" />
         <item itemvalue="TestNG.TestsRequest.getProductsWById" />
         <item itemvalue="TestNG.TestsRequest.getProductsByNonexistentId" />
-        <item itemvalue="TestNG.TestsRequest.getAllUsers" />
       </list>
     </recent_temporary>
   </component>
@@ -175,8 +188,12 @@
       <option name="number" value="Default" />
       <option name="presentableId" value="Default" />
       <updated>1696973156786</updated>
+      <workItem from="1701532173052" duration="2171000" />
     </task>
     <servers />
+  </component>
+  <component name="TypeScriptGeneratedFilesManager">
+    <option name="version" value="3" />
   </component>
   <component name="Vcs.Log.Tabs.Properties">
     <option name="TAB_STATES">
